Remove historical export code and combine base and streaming
